LR4/pyproject.toml
[tool.poetry]
name = "lab4"
version = "0.1.0"
description = ""
authors = ["Levkov Gleb <Glebchanskiy@yandex.by>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
poetry = "^1.4.0"
sqlalchemy = "^2.0.4"
psycopg2 = "^2.9.5"
tabulate = "^0.9.0"
kivy = "^2.1.0"
kivymd = "^1.1.1"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "lab4.start_cli:cli"
gui = "lab4.start_gui:gui"
